refactored code for better readability
git-svn-id: https://svn.code.sf.net/p/dosmid/code/trunk@80 850bf26b-01f9-4b63-b2ae-8b26f6eb5263

+++ midi.c
@@ -112,7 +112,7 @@ struct midi_chunk_t *midi_readchunk(FILE *fd) {
   return(res);
 }
 
-int midi_readhdr(FILE *fd, int *format, int *tracks, unsigned int *timeunitdiv, struct midi_chunkmap_t *chunklist, int maxchunks) {
+int midi_readhdr(FILE *fd, int *format, int *tracks, unsigned short *timeunitdiv, struct midi_chunkmap_t *chunklist, int maxchunks) {
   struct midi_chunk_t *chunk;
   unsigned char rmidbuff[12];
   /* test for RMID format and rewind if not found */
